pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "llm-eval-compass"
version = "0.1.0"
description = "A production-grade framework for evaluating, comparing, and selecting LLMs"
requires-python = ">=3.10"
authors = [{ name = "Rahul Goel" }]
dependencies = [
    "ragas==0.2.14",
    "deepeval==1.5.3",
    "langchain==0.3.25",
    "langchain-openai==0.3.18",
    "langchain-anthropic==0.3.15",
    "langchain-aws==0.2.17",
    "langchain-google-genai==2.1.4",
    "boto3==1.38.12",
    "openai==1.78.1",
    "anthropic==0.49.0",
    "tiktoken==0.9.0",
    "pyyaml==6.0.2",
    "pandas==2.2.3",
    "numpy==2.2.5",
    "jinja2==3.1.6",
    "python-dotenv==1.1.0",
    "mlflow==2.22.0",
    "wandb==0.19.11",
    "requests==2.32.3",
    "rich==14.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
